change protobuf as static library;
Former-commit-id: 537fce8952238bd032b415903e1c48df566e2e35 [formerly bcefe7807f9a55f1a775ba3d318c3647af11cfc1] [formerly 9ac1cea3d047acda56410bfc9a8f0d68135fe0e8 [formerly d0aec4e4a7be7286c5ba09e59b5453491a47c128]]
Former-commit-id: a186f3cb1b8cdd3f8d05912055b96a04738f32c1 [formerly b461c270341a93c8739faf5d68b136c03612424c]
Former-commit-id: 12351744d7f18ad12a77d646d5291c5c29a35b2a

+++ Frame/SDK/NetPlugin/AFCWebSocktServer.cpp
@@ -394,7 +394,7 @@ namespace ark
         head.SetMsgID(msg_id);
         head.SetUID(actor_rid);
 
-        int whole_len = EnCode(head, msg, msg_len, out_data);
+        size_t whole_len = EnCode(head, msg, msg_len, out_data);
         if (whole_len == msg_len + ARK_PKG_BASE_HEAD::ARK_CS_HEADER_LENGTH)
         {
             return SendMsgToAllClient(out_data.c_str(), out_data.length());
